Fix return code type
(cherry picked from commit 770b75d9fdf9410bca8647a4c20e029fa8ac72f1)

+++ src/features/perform_privacy_operation/cmd_perform_privacy_operation.c
@@ -127,5 +127,5 @@ end:
     explicit_bzero(privateKeyDataSwapped, sizeof(privateKeyDataSwapped));
     explicit_bzero(&privateKey, sizeof(privateKey));
     explicit_bzero(privateKeyData, sizeof(privateKeyData));
-    return error;
+    return (uint16_t) error;
 }
